backend/requirements.txt
fastapi~=0.128.0
uvicorn~=0.40.0
sqlmodel~=0.0.31
pydantic~=2.12.5
aiosqlite~=0.22.1
sqlalchemy~=2.0.45
psutil~=7.2.1
pyautogui~=0.9.54
python-multipart~=0.0.21
requests~=2.32.5
httpx~=0.28.1
aiohttp~=3.13.3
aiofiles~=25.1.0
loguru~=0.7.3
pydantic-settings~=2.12.0
python-dotenv~=1.2.1
openai~=2.14.0
anthropic~=0.45.0
google-genai~=1.56.0
mcp~=1.25.0
# STT/TTS
edge-tts~=7.2.7
faster-whisper~=1.2.1
torch>=2.0.0
safetensors~=0.7.0
numpy~=2.2.6
sentence-transformers~=5.2.0
sounddevice~=0.5.3
soundfile~=0.13.1
# Image/Vision
pillow~=12.1.0
opencv-python~=4.12.0.88
easyocr~=1.7.2
python-bidi~=0.6.7
# Others
pytz~=2025.2
pyyaml~=6.0.3
beautifulsoup4~=4.14.3
pypdf~=3.17.0
python-docx~=1.2.0
pyperclip~=1.11.0
praat-parselmouth~=0.4.7
websockets~=15.0.1
tenacity~=9.1.2
tiktoken~=0.12.0
pydub~=0.25.1
jinja2>=3.1.6
markupsafe>=3.0.3
wasmtime>=26.0.0
APScheduler~=3.10.4
dateparser~=1.2.0
ruff
black
